0.9.5.3

+++ selfdrive/test/setup_device_ci.sh
@@ -29,6 +29,7 @@ sudo abctl --set_success
 
 # patch sshd config
 sudo mount -o rw,remount /
+echo tici-$(cat /proc/cmdline | sed -e 's/^.*androidboot.serialno=//' -e 's/ .*$//') | sudo tee /etc/hostname
 sudo sed -i "s,/data/params/d/GithubSshKeys,/usr/comma/setup_keys," /etc/ssh/sshd_config
 sudo systemctl daemon-reload
 sudo systemctl restart ssh
@@ -70,7 +71,6 @@ safe_checkout() {
   git checkout $GIT_COMMIT
   git clean -xdff
   git submodule sync
-  git submodule foreach --recursive "git reset --hard && git clean -xdff"
   git submodule update --init --recursive
   git submodule foreach --recursive "git reset --hard && git clean -xdff"
 
@@ -80,7 +80,9 @@ safe_checkout() {
   echo "git checkout done, t=$SECONDS"
   du -hs $SOURCE_DIR $SOURCE_DIR/.git
 
-  rsync -a --delete $SOURCE_DIR $TEST_DIR
+  if [ -z "SKIP_COPY" ]; then
+    rsync -a --delete $SOURCE_DIR $TEST_DIR
+  fi
 }
 
 unsafe_checkout() {
@@ -96,7 +98,6 @@ unsafe_checkout() {
   git reset --hard $GIT_COMMIT
   git clean -df
   git submodule sync
-  git submodule foreach --recursive "git reset --hard && git clean -df"
   git submodule update --init --recursive
   git submodule foreach --recursive "git reset --hard && git clean -df"
